Rename ProfitChart component file
