+++ ProfitChart.jsx
